My implementation (Python 2.7)
Might be the wrong version of python but it should work if you fix the scikitimage import statements. What I ended up writing became similar to your code independently although I use a marching squares algorithm and then find the largest polygon so this should help make it more robust to choosing the largest shape and less dependent on the specificity of thresholding to only select pixels from the box. 

This could easily be adapted such that all RGB images are sent to grayscale and the marching squares should be able to select for only the box regardless if other pixels make it through thresholding (making it robust to different box colors).